+++ NEWS.md
@@ -31,7 +31,7 @@ remotes::install_github("nickpoison/astsa/astsa_build")
 
 #### v2.0.1: (on GitHub) - small updates after v2.0:
 
-- added `SV.mle` to fit the classical model in Chapter 6. (uses `SVfilter` so maybe put that in the xBox ?? not sure yet)
+- added `SV.mle` to fit the classical model in Chapter 6. `SVfilter` is now part of `SV.mle` and the original script is now in the xBox as `xSVfilter`.
 
 - `sarima` - prettified the output